YaelFV: Revert previous 'fix' for flushBuffer error
The fix was actually something inside `yael` itself (avoiding calling
'.this.acquire' a second time) which was triggered when flushBuffer
was called more than once.  Now repeated calls to flushBuffer are OK,
so the buffer size could be reduced back to the 1000 it was originally,
but might as well leave it at 10000 since we have the resources.